snowFlake

+++ Zinterface/managerInterface.go
@@ -2,8 +2,8 @@ package Zinterface
 
 type ManagerI interface {
 	Add(i ConnectionI)error
-	Get(id uint32)(ConnectionI,error)
-	Remove(id uint32)error
+	Get(id uint64)(ConnectionI,error)
+	Remove(id uint64)error
 	Size()int
 	Clear()error
 }
